Divide block_utils.hpp into separate files
